requirements.txt
Flask
flask-cors
pyyaml
sqlalchemy
elasticsearch
requests
pymysql
